Refactor endpoint factory registration and lookup
